+++ .gitignore
@@ -1,5 +1,6 @@
 # Built files
 /*.js
+!jest.config.js
 /*.map
 /*.d.ts
 native/artifacts.json
